Download samples from GitHub releases instead of installing package
This change downloads the samples from the GitHub releases instead of
installing the package. As of tag v11.6 all samples are accounted for
and for the time being we use this tag regardless of the CUDA image
version specified. The --wildcards argument is provided to tar to
extract only the required samples.

Signed-off-by: Evan Lezar <elezar@nvidia.com>